+++ jena-kafka-connector/src/main/java/org/apache/jena/kafka/KConnectorDesc.java
@@ -16,9 +16,12 @@
 
 package org.apache.jena.kafka;
 
+import java.time.Duration;
 import java.util.List;
 import java.util.Objects;
 import java.util.Properties;
+import java.util.function.Function;
+import java.util.function.Predicate;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -33,12 +36,21 @@ import org.apache.kafka.clients.consumer.ConsumerConfig;
  * href="https://github.com/telicent-oss/smart-caches-core/blob/main/docs/event-sources/kafka.md">Telicent Smart Cache
  * Core Libraries</a>.
  * </p>
+ * <p>
+ * Transaction management and handling is done by {@link org.apache.jena.kafka.common.FusekiProjector}, a number of
+ * methods on the connector provide advanced configuration tuning for this.  The values returned by these methods are
+ * driven from custom Fuseki Kafka configuration properties supplied as part of the Kafka configuration properties
+ * available via {@link #getKafkaConsumerProps()}.  See individual methods for what configuration properties to use, and
+ * the defaults used if not explicitly configured.
+ * </p>
  */
 @Getter
 @ToString
 @EqualsAndHashCode
 public class KConnectorDesc {
 
+    private static final Predicate<Integer> IS_POSITIVE_INTEGER = x -> x != null && x > 0;
+    private static final Predicate<Long> IS_POSITIVE_LONG = x -> x != null && x > 0;
     // Source
     private final List<String> topics;
     private final String dlqTopic;
@@ -95,18 +107,144 @@ public class KConnectorDesc {
         return this.kafkaConsumerProps.getProperty(ConsumerConfig.GROUP_ID_CONFIG);
     }
 
+    private <T> T fromKafkaProperties(String key, Function<String, T> parser, Predicate<T> validator, T defaultValue) {
+        String rawValue = this.kafkaConsumerProps.getProperty(key);
+        if (StringUtils.isBlank(rawValue)) {
+            return defaultValue;
+        } else {
+            try {
+                T value = parser.apply(rawValue);
+                return validator.test(value) ? value : defaultValue;
+            } catch (Throwable e) {
+                return defaultValue;
+            }
+        }
+    }
+
+    private <T> T fromKafkaProperties(String[] keys, Function<String, T> parser, Predicate<T> validator,
+                                      T defaultValue) {
+        for (String key : keys) {
+            T value = fromKafkaProperties(key, parser, validator, null);
+            if (value != null) {
+                return value;
+            }
+        }
+        return defaultValue;
+    }
+
     /**
      * Gets the maximum poll records as configured for the connector
      *
      * @return Maximum poll records
      */
     public int getMaxPollRecords() {
-        String rawValue = this.kafkaConsumerProps.getProperty(ConsumerConfig.MAX_POLL_RECORDS_CONFIG);
-        if (StringUtils.isBlank(rawValue)) {
-            return SysJenaKafka.KAFKA_FETCH_POLL_SIZE;
-        } else {
-            return Integer.parseInt(rawValue);
-        }
+        return fromKafkaProperties(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, Integer::parseInt, IS_POSITIVE_INTEGER,
+                                   SysJenaKafka.KAFKA_FETCH_POLL_SIZE);
+    }
+
+    /**
+     * Gets the batch size to use, in number of events, this is taken from the following in precedence order:
+     * <ol>
+     *     <li>The custom Fuseki Kafka configuration property {@value SysJenaKafka#FUSEKI_KAFKA_BATCH_SIZE}</li>
+     *     <li>The Kafka configuration property {@value ConsumerConfig#MAX_POLL_RECORDS_CONFIG}</li>
+     *     <li>The Fuseki Kafka default {@value SysJenaKafka#KAFKA_FETCH_POLL_SIZE}</li>
+     * </ol>
+     *
+     * @return Batch size, in number of events, to use
+     */
+    public int getBatchSize() {
+        return fromKafkaProperties(
+                new String[] { SysJenaKafka.FUSEKI_KAFKA_BATCH_SIZE, ConsumerConfig.MAX_POLL_RECORDS_CONFIG },
+                Integer::parseInt, IS_POSITIVE_INTEGER, SysJenaKafka.KAFKA_FETCH_POLL_SIZE);
+    }
+
+    /**
+     * Gets the batch size to use, in number of bytes, this is taken from the following in precedence order:
+     * <ol>
+     *     <li>The custom Fuseki Kafka configuration property {@value SysJenaKafka#FUSEKI_KAFKA_BATCH_SIZE_BYTES}</li>
+     *     <li>The Kafka configuration property {@value ConsumerConfig#FETCH_MAX_BYTES_CONFIG}</li>
+     *     <li>The Kafka configuration property {@value ConsumerConfig#MAX_PARTITION_FETCH_BYTES_CONFIG}</li>
+     *     <li>The Fuseki Kafka default {@value SysJenaKafka#DEFAULT_HIGH_LAG_BATCH_BYTE_THRESHOLD}</li>
+     * </ol>
+     * <p>
+     * Note that this configuration is only used when the {@link #getHighLagThreshold()} is exceeded.
+     * </p>
+     *
+     * @return Batch size, in number of bytes, to use
+     */
+    public long getBatchSizeBytes() {
+        return fromKafkaProperties(new String[] {
+                SysJenaKafka.FUSEKI_KAFKA_BATCH_SIZE_BYTES,
+                ConsumerConfig.FETCH_MAX_BYTES_CONFIG,
+                ConsumerConfig.MAX_PARTITION_FETCH_BYTES_CONFIG
+        }, Long::parseLong, IS_POSITIVE_LONG, SysJenaKafka.DEFAULT_HIGH_LAG_BATCH_BYTE_THRESHOLD);
+    }
+
+    /**
+     * Gets the number of batch sizes to track in order to average batch sizes over this and determine whether we're
+     * connected to low volume topics.
+     * <p>
+     * Configured by the custom Fuseki Kafka property {@value SysJenaKafka#FUSEKI_KAFKA_BATCH_SIZE_TRACKING}, using a
+     * default of {@value SysJenaKafka#DEFAULT_BATCH_SIZE_TRACKING_WINDOW}.  Once at least this many batches have been
+     * processed if the average batch size falls below the configured {@link #getLowVolumeBatchSizeThreshold()} then the
+     * {@link org.apache.jena.kafka.common.FusekiProjector} switches over to low volume batching mode.  In this mode it
+     * does not automatically commit as soon as it reaches lag of 0 instead waiting until either the batch size, or the
+     * {@link #getMaxTransactionDuration()} is exceeded.
+     * </p>
+     *
+     * @return Number of batch sizes to track
+     */
+    public int getBatchSizeTrackingWindow() {
+        return fromKafkaProperties(SysJenaKafka.FUSEKI_KAFKA_BATCH_SIZE_TRACKING, Integer::parseInt,
+                                   IS_POSITIVE_INTEGER,
+                                   SysJenaKafka.DEFAULT_BATCH_SIZE_TRACKING_WINDOW);
+    }
+
+    /**
+     * Gets the low volume batch size threshold
+     * <p>
+     * If the average batch size over the last {@link #getBatchSizeTrackingWindow()} batches is less than, or equal to,
+     * this threshold then the {@link org.apache.jena.kafka.common.FusekiProjector} switches over to low volume batching
+     * mode.  See {@link #getBatchSizeTrackingWindow()} for more details on this.
+     * </p>
+     *
+     * @return Low volume batch size threshold
+     */
+    public int getLowVolumeBatchSizeThreshold() {
+        return fromKafkaProperties(SysJenaKafka.FUSEKI_KAFKA_LOW_VOLUME_THRESHOLD, Integer::parseInt,
+                                   IS_POSITIVE_INTEGER,
+                                   SysJenaKafka.DEFAULT_AVERAGE_BATCH_SIZE_LOW_VOLUME_THRESHOLD);
+    }
+
+    /**
+     * Gets the high lag threshold, in number of events.
+     * <p>
+     * If the lag exceeds this then the {@link org.apache.jena.kafka.common.FusekiProjector} switches to high lag
+     * batching mode.  In this mode it no longer honours the {@link #getBatchSize()} instead only committing once
+     * {@link #getBatchSizeBytes()} has been reached.
+     * </p>
+     *
+     * @return High lag threshold
+     */
+    public long getHighLagThreshold() {
+        return fromKafkaProperties(SysJenaKafka.FUSEKI_KAFKA_HIGH_LAG_THRESHOLD, Long::parseLong,
+                                   IS_POSITIVE_LONG,
+                                   SysJenaKafka.DEFAULT_HIGH_LAG_THRESHOLD);
+    }
+
+    /**
+     * Gets the maximum transaction duration
+     * <p>
+     * This is expressed as an ISO 8601 Duration in the Kafka configuration e.g. {@code PT5M} is 5 minutes, defaults to
+     * {@link SysJenaKafka#DEFAULT_MAX_TRANSACTION_DURATION}.
+     * </p>
+     *
+     * @return Max transaction duration
+     */
+    public Duration getMaxTransactionDuration() {
+        return fromKafkaProperties(SysJenaKafka.FUSEKI_KAFKA_MAX_TRANSACTION_DURATION, Duration::parse,
+                                   SysJenaKafka::isValidDuration,
+                                   SysJenaKafka.DEFAULT_MAX_TRANSACTION_DURATION);
     }
 
 }